save ckpt every n steps (#428)
* save_every arg

* rename

* handle none case

---------

+++ src/zeroband/train.py
@@ -479,7 +479,10 @@ def train(config: TrainingConfig):
 
                 if len(previous_ckpt_rollout) > config.async_level:
                     path_to_delete = previous_ckpt_rollout.pop(0)
-                    if path_to_delete.exists():
+                    ckpt_step = int(str(path_to_delete).split("_")[-1])
+
+                    should_keep = config.ckpt.interval_rollout is not None and ckpt_step % config.ckpt.interval_rollout == 0
+                    if path_to_delete.exists() and not should_keep:
                         logger.info(f"Removing past rollout ckpt at {path_to_delete}")
                         shutil.rmtree(path_to_delete, ignore_errors=True)
